requirements.txt
python-dotenv
langchain
langchain-ollama
langchain-community
beautifulsoup4
streamlit